Switch styles to MaterialWPF
