+++ src/PlatformInstaller.Tests/packages.config
@@ -1,12 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ApprovalTests" version="3.0.11" targetFramework="net45" />
-  <package id="ApprovalUtilities" version="3.0.11" targetFramework="net45" />
-  <package id="Autofac" version="3.5.2" targetFramework="net45" />
-  <package id="Caliburn.Micro" version="3.0.1" targetFramework="net45" />
-  <package id="Caliburn.Micro.Core" version="3.0.1" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
-  <package id="NUnit" version="3.4.1" targetFramework="net45" />
-  <package id="ObjectApproval" version="1.4.0" targetFramework="net45" />
-  <package id="Serilog" version="2.0.0" targetFramework="net45" />
+  <package id="ApprovalTests" version="3.0.13" targetFramework="net45" />
+  <package id="ApprovalUtilities" version="3.0.13" targetFramework="net45" />
+  <package id="Autofac" version="4.4.0" targetFramework="net45" />
+  <package id="Caliburn.Micro" version="3.0.3" targetFramework="net45" />
+  <package id="Caliburn.Micro.Core" version="3.0.3" targetFramework="net45" />
+  <package id="MaterialDesignColors" version="1.1.3" targetFramework="net45" />
+  <package id="MaterialDesignThemes" version="2.3.0.823" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net45" />
+  <package id="NUnit" version="3.6.1" targetFramework="net45" />
+  <package id="ObjectApproval" version="1.6.0" targetFramework="net45" />
+  <package id="Serilog" version="2.4.0" targetFramework="net45" />
 </packages>
