+++ src/PlatformInstaller.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30110.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlatformInstaller", "PlatformInstaller\PlatformInstaller.csproj", "{3F4B1E65-2BAE-48C4-AD4C-5AD4E3D4EB34}"
 EndProject
@@ -44,11 +44,13 @@ Global
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Debug|x86.Build.0 = Debug|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|x86.ActiveCfg = Release|Any CPU
+		{E49C739F-3B27-4D62-AE16-04C155C1ED20}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
