--- src/PlatformInstaller/ContainerFactory.cs
@@ -1,63 +1,0 @@
-﻿using System.Linq;
-using System.Reflection;
-using Autofac;
-using Caliburn.Micro;
-using Mindscape.Raygun4Net;
-
-public static class ContainerFactory
-{
-    public static IContainer BuildContainer()
-    {
-        var builder = new ContainerBuilder();
-        builder.Register<IWindowManager>(c => new WindowManager()).InstancePerLifetimeScope();
-        builder.RegisterAssemblyTypes(ThisAssembly())
-            .Where(ViewModelConventions.IsInstanceViewOrModel)
-            .AsSelf()
-            .InstancePerDependency();
-        builder.RegisterType<EventAggregator>()
-            .As<IEventAggregator>()
-            .SingleInstance();
-        builder.RegisterType<WindowManager>()
-            .As<IWindowManager>()
-            .SingleInstance();
-        builder.RegisterType<RebootMachine>()
-            .SingleInstance();
-        builder.RegisterType<AbortInstallationHandler>()
-            .SingleInstance();
-        builder.RegisterType<InstallFeedbackReporter>()
-            .SingleInstance();
-        builder.RegisterType<ShellViewModel>()
-            .SingleInstance();
-        builder.RegisterType<ProcessRunner>()
-            .SingleInstance();
-        builder.RegisterType<CredentialStore>()
-            .SingleInstance();
-        builder.RegisterType<ProxyTester>()
-            .SingleInstance();
-        builder.RegisterType<ReleaseManager>()
-            .SingleInstance();
-        builder.RegisterType<Installer>()
-            .SingleInstance();
-        builder.RegisterType<LicenseAgreement>()
-            .SingleInstance();
-        builder.RegisterType<AppBootstrapper>()
-            .SingleInstance();
-        builder.RegisterType<AutoSubscriber>();
-        builder.RegisterType<PendingRestartAndResume>()
-            .SingleInstance();
-        builder.RegisterAssemblyTypes(ThisAssembly())
-            .Where(t => t.GetInterfaces()
-            .Any(i => i.IsAssignableFrom(typeof(IInstaller))))
-            .SingleInstance().AsImplementedInterfaces();
-        builder.RegisterInstance(new RaygunClient(Program.RaygunApiKey))
-            .SingleInstance();
-        builder.RegisterType<RuntimeUpgradeManager>()
-            .SingleInstance();
-        return builder.Build();
-    }
-
-    static Assembly ThisAssembly()
-    {
-        return typeof(ContainerFactory).Assembly;
-    }
-}
