--- src/PlatformInstaller/Controls/Footer.xaml
@@ -1,29 +1,0 @@
-﻿<UserControl x:Class="Footer"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:global="clr-namespace:"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../Styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="Auto" />
-        </Grid.ColumnDefinitions>
-        <Border Grid.ColumnSpan="3" BorderBrush="{StaticResource Gray90Brush}" BorderThickness="0,1,0,0" Margin="-20,5,-20, 5" Padding="20"/>
-        <TextBlock Grid.Column="0" Margin="20,0,0,0"  VerticalAlignment="Center">
-            <global:Link NavigateUri="http://docs.particular.net/platform/installer/" Style="{StaticResource {x:Type Hyperlink}}">Help</global:Link>
-        </TextBlock>
-        <TextBlock x:Name="AppVersion" Grid.Column="2" Foreground="{StaticResource Gray20Brush}"  Margin="0,0,20,0" VerticalAlignment="Center" HorizontalAlignment="Right" />
-    </Grid>
-</UserControl>
-
