--- src/PlatformInstaller/Controls/Footer.xaml.cs
@@ -1,10 +1,0 @@
-﻿using System.Reflection;
-
-public partial class Footer
-{
-    public Footer()
-    {
-        InitializeComponent();
-        AppVersion.Text = $"Version: {Assembly.GetExecutingAssembly().GetName().Version}";
-    }
-}
