+++ src/PlatformInstaller/DotNetUpgrade/Views/DotNetDownloadView.xaml
@@ -6,15 +6,8 @@
     <UserControl.DataContext>
         <global:DotNetDownloadViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../../Styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel>
-        <TextBlock DockPanel.Dock="Top" FontSize="18" FontWeight="Bold">Downloading .NET Framework 4.5.2</TextBlock>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource MaterialDesignTitleTextBlock}" FontWeight="Bold">Downloading .NET Framework 4.5.2</TextBlock>
         <Grid DockPanel.Dock="Top" Margin="0 10 0 0">
             <ProgressBar Height="20"  Minimum="0" Maximum="100" Value="{Binding PercentComplete}" Foreground="#2EA3C4"  />
             <TextBlock Text="{Binding Description}" VerticalAlignment="Center" HorizontalAlignment="Center" />
