--- src/PlatformInstaller/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-class Program
-{
-    public const string RaygunApiKey = "FOsaSSOiJ59Np6IQKHx4Kg==";
-
-    [STAThread]
-    public static void Main()
-    {
-        Runner.Run();
-    }
-}
