--- src/PlatformInstaller/Styles.xaml
@@ -1,32 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="/Styles/Palette.xaml"/>
-        <ResourceDictionary Source="/Styles/Button.xaml" />
-        <ResourceDictionary Source="/Styles/Hyperlink.xaml"/>
-    </ResourceDictionary.MergedDictionaries>
- <!--
-    <Style TargetType="{x:Type Button}">
-        <Setter Property="Foreground" Value="White"/>
-        <Setter Property="Margin" Value="5"/>
-        <Setter Property="Padding" Value="10,0,10,0"/>
-
-        <Style.Triggers>
-            <Trigger Property="IsEnabled" Value="true">
-                <Setter Property="BorderBrush" Value="#1D729B" />
-                <Setter Property="Background" Value="#2EA3C4" />
-            </Trigger>
-        </Style.Triggers>
-    </Style>
-   -->
-    <Style TargetType="{x:Type TextBlock}">
-        <Setter Property="TextWrapping" Value="Wrap"/>
-    </Style>
-    <Style TargetType="{x:Type TextBox}">
-        <Setter Property="TextWrapping" Value="Wrap"/>
-        <Setter Property="IsReadOnly" Value="True"/>
-    </Style>
-    <Style TargetType="{x:Type CheckBox}">
-        <Setter Property="Foreground" Value="Blue"/>
-    </Style>
-</ResourceDictionary>
