--- src/PlatformInstaller/Styles/Hyperlink.xaml
@@ -1,22 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="Palette.xaml"/>
-    </ResourceDictionary.MergedDictionaries>
-    
-    <Style TargetType="{x:Type Hyperlink}">
-        <Style.Triggers>
-            <Trigger Property="IsMouseOver" Value="true">
-                <Setter Property="Foreground" Value="{StaticResource LightThemeBrush}" />
-            </Trigger>
-            <Trigger Property="IsEnabled" Value="false">
-                <Setter Property="Foreground" Value="{StaticResource Gray60Brush}" />
-            </Trigger>
-            <Trigger Property="IsEnabled" Value="true">
-                <Setter Property="Cursor" Value="Hand" />
-            </Trigger>
-        </Style.Triggers>
-        <Setter Property="Foreground" Value="{StaticResource ThemeBrush}" />
-        <Setter Property="TextDecorations" Value="{x:Null}" />
-    </Style>
-</ResourceDictionary>
