--- src/PlatformInstaller/Styles/Palette.xaml
@@ -1,60 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-
-    <Color x:Key="BlackColor">#000000</Color>
-    <Color x:Key="Gray10Color">#1A1A1A</Color>
-    <Color x:Key="Gray20Color">#333333</Color>
-    <Color x:Key="Gray30Color">#4C4C4C</Color>
-    <Color x:Key="Gray40Color">#666666</Color>
-    <Color x:Key="Gray50Color">#808080</Color>
-    <Color x:Key="Gray60Color">#999999</Color>
-    <Color x:Key="Gray70Color">#B2B2B2</Color>
-    <Color x:Key="Gray80Color">#CCCCCC</Color>
-    <Color x:Key="Gray90Color">#E6E6E6</Color>
-    <Color x:Key="WhiteColor">#FFFFFF</Color>
-
-    <Color x:Key="LightThemeColor">#00BFE6</Color>
-    <Color x:Key="ThemeColor">#009FBF</Color>
-    <Color x:Key="DarkThemeColor">#008099</Color>
-
-    <Color x:Key="LightErrorColor">#FF0D4D</Color>
-    <Color x:Key="ErrorColor">#E6003D</Color>
-    <Color x:Key="DarkErrorColor">#BF0033</Color>
-
-    <Color x:Key="SuccessColor">#3DB982</Color>
-
-    <Color x:Key="LightWarningColor">#FFfD0D</Color>
-    <Color x:Key="WarningColor">#E6aD00</Color>
-    <Color x:Key="DarkWarningColor">#BF7d00</Color>
-
-
-    <SolidColorBrush x:Key="BlackBrush" Color="{StaticResource BlackColor}" />
-    <SolidColorBrush x:Key="Gray10Brush" Color="{StaticResource Gray10Color}" />
-    <SolidColorBrush x:Key="Gray20Brush" Color="{StaticResource Gray20Color}" />
-    <SolidColorBrush x:Key="Gray30Brush" Color="{StaticResource Gray30Color}" />
-    <SolidColorBrush x:Key="Gray40Brush" Color="{StaticResource Gray40Color}" />
-    <SolidColorBrush x:Key="Gray50Brush" Color="{StaticResource Gray50Color}" />
-    <SolidColorBrush x:Key="Gray60Brush" Color="{StaticResource Gray60Color}" />
-    <SolidColorBrush x:Key="Gray70Brush" Color="{StaticResource Gray70Color}" />
-    <SolidColorBrush x:Key="Gray80Brush" Color="{StaticResource Gray80Color}" />
-    <SolidColorBrush x:Key="Gray90Brush" Color="{StaticResource Gray90Color}" />
-    <SolidColorBrush x:Key="WhiteBrush" Color="{StaticResource WhiteColor}" />
-
-    <SolidColorBrush x:Key="LightThemeBrush" Color="{StaticResource LightThemeColor}" />
-    <SolidColorBrush x:Key="ThemeBrush" Color="{StaticResource ThemeColor}" />
-    <SolidColorBrush x:Key="DarkThemeBrush" Color="{StaticResource DarkThemeColor}" />
-
-    <SolidColorBrush x:Key="LightErrorBrush" Color="{StaticResource LightErrorColor}" />
-    <SolidColorBrush x:Key="ErrorBrush" Color="{StaticResource ErrorColor}" />
-    <SolidColorBrush x:Key="DarkErrorBrush" Color="{StaticResource DarkErrorColor}" />
-
-    <SolidColorBrush x:Key="SuccessBrush" Color="{StaticResource SuccessColor}" />
-
-    <SolidColorBrush x:Key="LightWarningBrush" Color="{StaticResource LightWarningColor}" />
-    <SolidColorBrush x:Key="WarningBrush" Color="{StaticResource WarningColor}" />
-    <SolidColorBrush x:Key="DarkWarningBrush" Color="{StaticResource DarkWarningColor}" />
-
-
-    <SolidColorBrush x:Key="LightBackgroundBrush" Color="{StaticResource Gray80Color}" />
-    <SolidColorBrush x:Key="BorderBrush" Color="{StaticResource Gray60Color}" />
-    <SolidColorBrush x:Key="ForegroundBrush" Color="{StaticResource Gray20Color}" />
-</ResourceDictionary>
