+++ src/PlatformInstaller/Views/ConfirmAbortInstallView.xaml
@@ -2,22 +2,16 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         x:Class="ConfirmAbortInstallView"  WindowStartupLocation="CenterScreen"
-        Title="Particular Platform Installer" Height="190" Width="400" WindowStyle="None" ResizeMode="NoResize" BorderBrush="White" >
-    <Window.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Window.Resources>
+        Height="190" Width="400" 
+        Style="{StaticResource MaterialDesignWindowStyle}"
+    >
     <Border BorderBrush="#20000000" BorderThickness="2">
         <StackPanel Margin="10" VerticalAlignment="Center">
-
-            <TextBlock Margin="0,10,0,0" FontSize="15" FontWeight="DemiBold" >Exit Confirmation</TextBlock>
-            <TextBlock Margin="0,10,0,0" FontSize="13" TextWrapping="Wrap">Are you sure you wish to close the installer while an installation running? Some components may be left in an unknown state.</TextBlock>
+            <TextBlock Margin="0,10,0,0" Style="{StaticResource MaterialDesignTitleTextBlock}">Exit Confirmation</TextBlock>
+            <TextBlock Margin="0,10,0,0" Style="{StaticResource MaterialDesignBody1TextBlock}" TextWrapping="Wrap">Are you sure you wish to close the installer while an installation running? Some components may be left in an unknown state.</TextBlock>
             <StackPanel Orientation="Horizontal" Margin="0,10,0,0"  FlowDirection="RightToLeft">
                 <Button x:Name="ConfirmCancellation">Exit</Button>
-                <Button x:Name="ContinueInstallation" Style="{StaticResource HiliteButton}">Back</Button>
+                <Button x:Name="ContinueInstallation" Margin="10,0">Back</Button>
             </StackPanel>
         </StackPanel>
     </Border>
