+++ src/PlatformInstaller/Views/FailedInstallationView.xaml
@@ -10,20 +10,15 @@
     <UserControl.DataContext>
         <global:FailedInstallationViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel >
-        <TextBlock DockPanel.Dock="Top" FontSize="14" FontWeight="DemiBold" Text="{Binding FailureDescription}"/>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource MaterialDesignSubheadingTextBlock}" Text="{Binding FailureDescription}" TextWrapping="Wrap"/>
         <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" FlowDirection="RightToLeft">
             <Button x:Name="Exit" Content="Exit"/>
-            <Button x:Name="Home" Content="Back to Main"/>
-            <Button x:Name="ReportError" Style="{StaticResource HiliteButton}" Content="Report Error" />
+            <Button x:Name="Home" Content="Back to Main"  Margin="10,0"/>
+            <Button x:Name="ReportError" Content="Report Error" />
         </StackPanel>
-        <TextBox DockPanel.Dock="Top" HorizontalScrollBarVisibility="Visible" Margin="0 10 0 0" VerticalScrollBarVisibility="Visible" IsReadOnly="True" Text="{Binding FailuresText}" Foreground="#2EA3C4" FontSize="10" FontWeight="Medium"/>
+        <TextBox DockPanel.Dock="Top" HorizontalScrollBarVisibility="Visible"
+                 Margin="0 10 0 0" VerticalScrollBarVisibility="Visible" 
+                 IsReadOnly="True" Text="{Binding FailuresText}" />
     </DockPanel>
 </UserControl>
