+++ src/PlatformInstaller/Views/FailureView.xaml
@@ -9,19 +9,14 @@
     <UserControl.DataContext>
         <global:FailureViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel>
-        <TextBlock DockPanel.Dock="Top" FontSize="14" FontWeight="DemiBold" Text="{Binding FailureDescription}"/>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource MaterialDesignSubheadingTextBlock}" Text="{Binding FailureDescription}" TextWrapping="Wrap"/>
         <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" FlowDirection="RightToLeft">
             <Button x:Name="Exit" Content="Exit"/>
-            <Button x:Name="ReportError" Style="{StaticResource HiliteButton}" Content="Report Error" />
+            <Button x:Name="ReportError"  Content="Report Error" Margin="10,0"/>
         </StackPanel>
-        <TextBox DockPanel.Dock="Top" HorizontalScrollBarVisibility="Visible" Margin="0 10 0 0" VerticalScrollBarVisibility="Visible" IsReadOnly="True" Text="{Binding FailureText}" Foreground="#2EA3C4" FontSize="10" FontWeight="Medium"/>
+        <ScrollViewer>
+            <TextBlock DockPanel.Dock="Top" Margin="0 10 0 0" Text="{Binding FailureText}" TextWrapping="Wrap"/>    
+        </ScrollViewer>
     </DockPanel>
 </UserControl>
