+++ src/PlatformInstaller/Views/InstallingView.xaml
@@ -10,14 +10,6 @@
     <UserControl.DataContext>
         <global:InstallingViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-        <BooleanToVisibilityConverter x:Key="BoolToVis" />
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel>
         <Grid DockPanel.Dock="Top">
             <Grid.RowDefinitions>
@@ -26,13 +18,13 @@
                 <RowDefinition Height="Auto"/>
                 <RowDefinition Height="Auto"/>
             </Grid.RowDefinitions>
-            <TextBlock Grid.Row="0" FontSize="18" FontWeight="DemiBold">Installing the Particular Service Platform</TextBlock>
-            <TextBlock Grid.Row="1" Foreground="{StaticResource Gray30Brush}">This may take a few minutes</TextBlock>
+            <TextBlock Grid.Row="0" Style="{StaticResource MaterialDesignTitleTextBlock}">Installing the Particular Service Platform</TextBlock>
+            <TextBlock Grid.Row="1" Style="{StaticResource MaterialDesignSubheadingTextBlock}">This may take a few minutes</TextBlock>
             <StackPanel Grid.Row="2" Orientation="Horizontal" Margin="0 5 0 0">
                 <TextBlock FontSize="12" FontWeight="DemiBold" Margin="0 0 10 0" >Installing:</TextBlock>
                 <TextBlock Text="{Binding CurrentStatus}" />
             </StackPanel>
-            <ProgressBar Grid.Row="3" Height="20" Margin="0 10 0 0" Minimum="0" Maximum="{Binding InstallCount}" Value="{Binding InstallProgress}" Foreground="#2EA3C4"/>
+            <ProgressBar Grid.Row="3" Height="20" Margin="0 10 0 0" Minimum="0" Maximum="{Binding InstallCount}" Value="{Binding InstallProgress}" />
         </Grid>
 
         <Grid DockPanel.Dock="Top" Visibility="{Binding Path=Downloading, Converter={StaticResource BoolToVis}}">
@@ -44,7 +36,7 @@
                 <TextBlock FontSize="12" FontWeight="Bold" Margin="0 0 10 0" >Download Progress:</TextBlock>
                 <TextBlock Text="{Binding NestedActionDescription}" />
             </StackPanel>
-            <ProgressBar Grid.Row="1" Height="20" Margin="0 10 0 0" Minimum="0" Maximum="100" Value="{Binding NestedActionPercentComplete}" Foreground="#2EA3C4"/>
+            <ProgressBar Grid.Row="1" Height="20" Margin="0 10 0 0" Minimum="0" Maximum="100" Value="{Binding NestedActionPercentComplete}"/>
         </Grid>
 
         <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" FlowDirection="RightToLeft">
