+++ src/PlatformInstaller/Views/LicenseAgreementView.xaml
@@ -9,22 +9,15 @@
     <UserControl.DataContext>
         <global:LicenseAgreementViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel>
         <StackPanel DockPanel.Dock="Top" Orientation="Vertical" FlowDirection="LeftToRight">
-            <TextBlock FontSize="16" Margin="6">
+            <TextBlock  Margin="6"  Style="{StaticResource  MaterialDesignSubheadingTextBlock}" TextWrapping="Wrap">
                 Your initial trial lasts for 14 days. When the trial expires we'll provide instructions on how to extend it for <Bold>free</Bold>.
             </TextBlock>
         </StackPanel>
         <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" FlowDirection="RightToLeft" >
             <Button x:Name="Exit" Content="Exit"/>
-            <Button x:Name="Agree" Content="I Agree"/>
+            <Button x:Name="Agree" Content="I Agree" Margin="10,0"/>
         </StackPanel>
         <global:BindableRichText Text="{Binding License, Mode=OneTime}" IsReadOnly="True" VerticalScrollBarVisibility="Visible" Margin="0,10"/>
     </DockPanel>
