+++ src/PlatformInstaller/Views/ProxySettingsView.xaml
@@ -1,27 +1,62 @@
 ﻿<Window
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
         x:Class="ProxySettingsView"  WindowStartupLocation="CenterScreen"
-        Title="Particular Platform Installer" Height="275" Width="300" WindowStyle="None" ResizeMode="NoResize" BorderBrush="White" >
+        Height="300" Width="300"
+        Style="{StaticResource MaterialDesignWindowStyle}"
+    >
     <Window.Resources>
         <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
+            <Style TargetType="{x:Type TextBox}" BasedOn="{StaticResource MaterialDesignTextBox}">
+                <Setter Property="Margin" Value="0 8 0 8" />
+            </Style>
+            <Style TargetType="{x:Type PasswordBox}" BasedOn="{StaticResource MaterialDesignPasswordBox}">
+                <Setter Property="Margin" Value="0 8 0 8" />
+            </Style>
+            </ResourceDictionary>
     </Window.Resources>
     <Border BorderBrush="#20000000" BorderThickness="2">
         <StackPanel Margin="10">
             <Image Source="pack://application:,,,/PlatformInstaller;component/Images/particularSmall.png" HorizontalAlignment="Left" Stretch="None" Margin="0,0,0,10"/>
-            <TextBlock Margin="0,10,0,0" DockPanel.Dock="Top" FontSize="18" FontWeight="DemiBold">Proxy Credentials required</TextBlock>
-            <Label Content="Login Name" Margin="-5,0,0,0"/>
-            <TextBox x:Name="username" VerticalAlignment="Center" IsReadOnly="False"/>
-            <Label Margin="-5,0,0,0" >Password</Label>
-            <PasswordBox x:Name="password" VerticalAlignment="Center"/>
+            <TextBlock Margin="0,10,0,0" DockPanel.Dock="Top" Style="{StaticResource MaterialDesignSubheadingTextBlock}" >Proxy Credentials required</TextBlock>
+            <Grid>
+                <Grid.RowDefinitions>
+                    <RowDefinition  Height="Auto"/>
+                    <RowDefinition  Height="Auto"/>
+                    <RowDefinition  Height="Auto"/>
+                    <RowDefinition  Height="Auto"/>
+                    <RowDefinition  Height="Auto"/>
+                </Grid.RowDefinitions>
+
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="30"/>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="10"/>
+                </Grid.ColumnDefinitions>
+
+                
+                <materialDesign:PackIcon Grid.Row="0" Grid.Column="0" Kind="Account"
+                                 Margin="0 12 0 0"
+                                 Foreground="{Binding ElementName=username, Path=BorderBrush}" />
+                <TextBox Grid.Row="0" Grid.Column="1" 		
+                     x:Name="username"
+                     Margin="0 12 0 0"
+                     materialDesign:HintAssist.Hint="Account Name" Style="{StaticResource MaterialDesignFloatingHintTextBox}" />
+                
+                <materialDesign:PackIcon Grid.Row="1" Grid.Column="0" Kind="Key"
+                                 Margin="0 12 0 0"
+                                 Foreground="{Binding ElementName=password, Path=BorderBrush}" />
+                <PasswordBox Grid.Row="1" Grid.Column="1" 		
+                     x:Name="password"
+                     Margin="0 12 0 0"
+                     materialDesign:HintAssist.Hint="Password" Style="{StaticResource MaterialDesignFloatingHintPasswordBox}" />
+
+            </Grid>
             <CheckBox x:Name="saveCredentials"  Content="Save Credentials" HorizontalAlignment="Center" Margin="0,10,0,9" Foreground="Black"/> 
             <StackPanel Orientation="Horizontal" Margin="0,10,0,0"  FlowDirection="RightToLeft">
                 <Button Click="CancelClick">Cancel</Button>
-                <Button Click="OkClick" Style="{StaticResource HiliteButton}" Padding="40,0">OK</Button>
+                <Button Click="OkClick"  Margin="10,0">OK</Button>
             </StackPanel>
         </StackPanel>
     </Border>
