+++ src/PlatformInstaller/Views/RebootNeededView.xaml
@@ -6,19 +6,12 @@
     <UserControl.DataContext>
         <global:RebootNeededViewModel/>
     </UserControl.DataContext>
-    <UserControl.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </UserControl.Resources>
     <DockPanel >
-        <TextBlock DockPanel.Dock="Top" FontSize="18" FontWeight="Bold">A reboot is needed!</TextBlock>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource MaterialDesignTitleTextBlock}">A reboot is needed!</TextBlock>
         <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" FlowDirection="RightToLeft">
             <Button x:Name="Exit"  Content="Exit"/>
-            <Button x:Name="Reboot" Style="{StaticResource HiliteButton}" Content="Reboot Now"/>
+            <Button x:Name="Reboot" Content="Reboot Now"  Margin="10,0"/>
         </StackPanel>
-        <TextBlock Foreground="Gray">Please reboot your machine and restart the platform installer</TextBlock>
+        <TextBlock Style="{StaticResource MaterialDesignSubheadingTextBlock}">Please reboot your machine and restart the platform installer</TextBlock>
     </DockPanel>
 </UserControl>
