+++ src/PlatformInstaller/Views/ResumeInstall.xaml
@@ -3,25 +3,20 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:global="clr-namespace:"
         x:Class="ResumeInstallView"  WindowStartupLocation="CenterScreen"
-        Title="Particular Platform Installer" Height="190" Width="400" WindowStyle="None" ResizeMode="NoResize" BorderBrush="White" >
+        Height="200" Width="400" WindowStyle="None"
+        Style="{StaticResource MaterialDesignWindowStyle}"
+    >
     <Window.DataContext>
         <global:ResumeInstallViewModel/>
     </Window.DataContext>
-    <Window.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Window.Resources>
     <Border BorderBrush="#20000000" BorderThickness="2">
         <StackPanel Margin="10">
             <Image Source="pack://application:,,,/PlatformInstaller;component/Images/particularSmall.png" HorizontalAlignment="Left" Stretch="None" />
-            <TextBlock Margin="0,10,0,0" FontSize="15" FontWeight="Bold" >Resume the previous installation?</TextBlock>
-            <TextBlock Margin="0,10,0,0" FontSize="13" TextWrapping="Wrap">The Platform Installer has completed the required reboot and may now resume the installation. Do you wish to continue?"</TextBlock>
+            <TextBlock Margin="0,10,0,0" Style="{StaticResource MaterialDesignTitleTextBlock}" >Resume the previous installation?</TextBlock>
+            <TextBlock Margin="0,10,0,0" TextWrapping="Wrap">The Platform Installer has completed the required reboot and may now resume the installation. Do you wish to continue?"</TextBlock>
             <StackPanel Orientation="Horizontal" Margin="0,10,0,0"  FlowDirection="RightToLeft">
-                <Button x:Name="ConfirmCancellation">Cancel</Button>
-                <Button x:Name="ContinueInstallation" Style="{StaticResource HiliteButton}">Continue installation</Button>
+                <Button x:Name="ConfirmCancellation" >Cancel</Button>
+                <Button x:Name="ContinueInstallation" Margin="10,0">Continue installation</Button>
             </StackPanel>
         </StackPanel>
     </Border>
