+++ src/PlatformInstaller/Views/ShellView.xaml
@@ -2,16 +2,11 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:global="clr-namespace:"
+        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
         x:Class="ShellView" WindowStartupLocation="CenterScreen"
-        Title="Particular Platform Installer" Height="600" Width="640" WindowStyle="None" ResizeMode="NoResize" BorderBrush="White" >
-    <Window.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="../styles.xaml"/>
-                <ResourceDictionary Source="../Styles/WindowCloseButton.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Window.Resources>
+        Height="600" Width="640" 
+        Style="{StaticResource MaterialDesignWindowStyle}"
+    >
     <Border BorderBrush="#20000000" BorderThickness="2">
         <Grid>
             <Grid.RowDefinitions>
@@ -22,20 +17,33 @@
             <StackPanel Grid.Row="0" Orientation="Horizontal" FlowDirection="RightToLeft" Margin="0" VerticalAlignment="Top" >
                     <Button x:Name="Exit"
                             Style="{StaticResource WindowCloseButtonStyle}"
-                            ToolTip="close">
-                            <ContentControl Width="16"
-                                Height="16"
-                                Foreground="Black"
-                                Template="{StaticResource CloseIcon}" />
+                            ToolTip="Close">
+                        <ContentControl Width="16" Height="16" Foreground="Black" Template="{StaticResource CloseIcon}" />
                     </Button>
 
-                    <TextBlock Margin="10,0,0,0" VerticalAlignment="Center">
-                        <Hyperlink Click="OpenLogDirectory">Logs</Hyperlink>
-                    </TextBlock>
+                <Button Style="{StaticResource MaterialDesignFlatButton}" HorizontalAlignment="Left" Click="OpenLogDirectory">
+                    <StackPanel Orientation="Horizontal">
+                        <TextBlock Margin="4,0" FontSize="12">Logs</TextBlock>
+                        <materialDesign:PackIcon Kind="Magnify" RenderSize="16,16" />
+                    </StackPanel>
+                </Button>
             </StackPanel>
             <Image Grid.Row="0" Source="pack://application:,,,/PlatformInstaller;component/Images/particularSmall.png" HorizontalAlignment="Left" Margin="20 20 20 0" Stretch="None" />
             <ContentControl Grid.Row="1" x:Name="ActiveItem" Margin="20,20,20,5" Focusable="False" />
-            <global:Footer Grid.Row="2"/>
+            <Grid Grid.Row="2">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <Border Grid.ColumnSpan="3" BorderThickness="0,1,0,0" Margin="-20,5,-20, 5" Padding="20" BorderBrush="{StaticResource MaterialDesignDivider}"/>
+                <Button Grid.Column="0" Style="{StaticResource MaterialDesignFlatButton}" FontSize="12" HorizontalAlignment="Left" Click="OpenDocs">
+                    <StackPanel Orientation="Horizontal">
+                        <TextBlock Margin="4,0">Help</TextBlock>
+                    </StackPanel>
+                </Button>
+                <TextBlock x:Name="AppVersion" Grid.Column="2" Style="{StaticResource MaterialDesignCaptionTextBlock}" Margin="0,0,20,0" VerticalAlignment="Center" HorizontalAlignment="Right" />
+            </Grid>
         </Grid>
     </Border>
 </Window>
