+++ src/PlatformInstaller/packages.config
@@ -1,19 +1,21 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Anotar.Serilog.Fody" version="3.1.0" targetFramework="net45" developmentDependency="true" />
+  <package id="Anotar.Serilog.Fody" version="3.2.0" targetFramework="net45" developmentDependency="true" />
   <package id="AsyncErrorHandler.Fody" version="1.0.11.0" targetFramework="net45" developmentDependency="true" />
-  <package id="Autofac" version="3.5.2" targetFramework="net45" userInstalled="true" />
-  <package id="Caliburn.Micro" version="3.0.1" targetFramework="net45" userInstalled="true" />
-  <package id="Caliburn.Micro.Core" version="3.0.1" targetFramework="net45" userInstalled="true" />
+  <package id="Autofac" version="4.4.0" targetFramework="net45" userInstalled="true" />
+  <package id="Caliburn.Micro" version="3.0.3" targetFramework="net45" userInstalled="true" />
+  <package id="Caliburn.Micro.Core" version="3.0.3" targetFramework="net45" userInstalled="true" />
   <package id="Caseless.Fody" version="1.4.2" targetFramework="net45" developmentDependency="true" />
   <package id="Fielder.Fody" version="1.0.12" targetFramework="net45" developmentDependency="true" />
   <package id="Fody" version="1.29.4" targetFramework="net45" developmentDependency="true" />
-  <package id="Janitor.Fody" version="1.2.1.0" targetFramework="net45" developmentDependency="true" />
-  <package id="Mindscape.Raygun4Net" version="5.3.1" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
-  <package id="PropertyChanged.Fody" version="1.51.3" targetFramework="net45" developmentDependency="true" />
-  <package id="Resourcer.Fody" version="1.4.2" targetFramework="net45" developmentDependency="true" />
-  <package id="Serilog" version="2.0.0" targetFramework="net45" />
-  <package id="Serilog.Sinks.File" version="2.1.0" targetFramework="net45" />
-  <package id="Serilog.Sinks.RollingFile" version="2.1.0" targetFramework="net45" />
+  <package id="Janitor.Fody" version="1.2.2.0" targetFramework="net45" developmentDependency="true" />
+  <package id="MaterialDesignColors" version="1.1.3" targetFramework="net45" />
+  <package id="MaterialDesignThemes" version="2.3.0.823" targetFramework="net45" />
+  <package id="Mindscape.Raygun4Net" version="5.4.1" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net45" />
+  <package id="PropertyChanged.Fody" version="1.52.1" targetFramework="net45" developmentDependency="true" />
+  <package id="Resourcer.Fody" version="1.4.3" targetFramework="net45" developmentDependency="true" />
+  <package id="Serilog" version="2.4.0" targetFramework="net45" />
+  <package id="Serilog.Sinks.File" version="3.2.0" targetFramework="net45" />
+  <package id="Serilog.Sinks.RollingFile" version="3.3.0" targetFramework="net45" />
 </packages>
