+++ src/Test-PlatformInstaller/Program.cs
@@ -5,6 +5,6 @@ class Program
     [STAThread]
     public static void Main()
     {
-        Runner.Run();
+       App.Main();
     }
 }
